[Refactor] 페르소나 의견 수집 시작 시 사용자 크레딧 차감

+++ src/main/java/kr/yuns/dropthepitchserver/credit/data/repository/OpinionRequestCreditRepository.java
@@ -3,5 +3,5 @@ package kr.yuns.dropthepitchserver.credit.data.repository;
 import kr.yuns.dropthepitchserver.credit.data.entity.OpinionRequestCredit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OpinionRequestCreditRepository extends JpaRepository<OpinionRequestCredit, Integer> {
+public interface OpinionRequestCreditRepository extends JpaRepository<OpinionRequestCredit, Long> {
 }
